+++ src/views/commit_list.rs
@@ -120,7 +120,10 @@ pub fn handle_key(action: KeyCommand, app: &mut AppState) -> AppAction {
             app.enter_move_select();
             AppAction::Handled
         }
-        KeyCommand::Mergetool | KeyCommand::None | KeyCommand::ForceQuit => AppAction::Handled,
+        KeyCommand::Mergetool
+        | KeyCommand::OpenEditor
+        | KeyCommand::None
+        | KeyCommand::ForceQuit => AppAction::Handled,
         KeyCommand::SeparatorLeft => {
             app.separator_offset = app.separator_offset.saturating_sub(4);
             AppAction::Handled